refactor(arona-core): 引入AronaRedis门面收敛Redis客户端访问(re.this迁移PR1)
把 redisDbQuery 的 receiver 从 KredsClient 收敛到项目自有窄门面 AronaRedis, 全代码库
(arona-core/kivotos/arona/user-recorder)调用点改走门面方法, 不再直接依赖具体 Redis 客户端类型。
门面吸收底层形态差异(hgetAll 扁平表 -> 保序 Map)、集中线格式取舍; 底层仍是 kreds, 行为不变,
为后续整体替换为 re.this 做准备。

+++ arona/arona-plugin/arona/src/main/kotlin/com/diyigemt/arona/arona/command/GachaCommand.kt
@@ -185,7 +185,7 @@ class GachaCommand : AbstractCommand(
             key,
             JsonIgnoreUnknownKeys.encodeToString(selfConfig)
           )
-          expire(key, (10u * 60u).toULong())
+          expire(key, 10L * 60L)
         }
         sendResult(pool.name, 10, result, code)
       }
